feat: add valueFilters, add render function goods, filters, goodsCount
