+++ src/assets/ts/lib/renderShopCards.ts
@@ -1,4 +1,5 @@
 import { store } from '../store';
+import { filterGoods, getFiltersFromUrl, getSearchStringFromUrl } from './filterGoods';
 import { getCardHtml } from './getCardHtml';
 
 export const renderShopCards = async (goodsRenderId: string) => {
@@ -7,7 +8,19 @@ export const renderShopCards = async (goodsRenderId: string) => {
     const view = store.view_settings.mode;
 
     const goodsCardsHtmlArr: HTMLElement[] = [];
-    console.log('render cards');
+
+    const filters = getFiltersFromUrl();
+
+    for (const filter in filters) {
+        if (Object.prototype.hasOwnProperty.call(filters, filter)) {
+            const element = filters[filter];
+            store.filters_settings[element.name] = element.value;
+        }
+    }
+
+    store.filteredGoodsItems = filterGoods(store.goodsItems, filters, getSearchStringFromUrl());
+
+    // console.log('render cards');
 
     for (let i = 0; i < store.filteredGoodsItems.length; i++)
         goodsCardsHtmlArr.push(await getCardHtml(store.filteredGoodsItems[i], { view }));
